Move external_timer implementations into separate detail headers
